requirements.txt
customtkinter==5.2.2
pillow==10.2.0
pyinstaller==6.11.1
